feat: holder 模块注册表移植 + tasker 并发修正(tsan 全绿)
- holder:FreeRTOS 信号量→pthread timedlock,依赖分批初始化/循环依赖置 ERROR 逻辑逐行保留
- 新增 test_holder:循环依赖、重复注册、ERROR 态隔离、必需模块停机、生命周期边界(61 断言)
- tasker:跨线程标志(done/cancel/dispatched/run_cnt/pri/level/stop/init_flag)统一 C11 原子访问,
  tasker_is_empty/is_full 补调度表锁——修复模板固有的 tsan 检出数据竞争,调度逻辑不变
- dg-test 7/7 常规全绿 + tsan 全绿;dg-build 零警告

+++ door-guard/tests/test_event_bus_stress.c
@@ -111,7 +111,7 @@ int main(void)
         pthread_join(th[t], NULL);
 
     /* 发布方全部返回并不代表分发完毕:等 processed 追平 published(上限 60s) */
-    const int32_t total = STRESS_THREADS * STRESS_PER_THREAD;
+    const uint32_t total = STRESS_THREADS * STRESS_PER_THREAD;
     uint32_t pub = 0, pro = 0, drop = 0, herr = 0;
     for (int i = 0; i < 6000; i++) {
         event_bus_get_stats(&pub, &pro, &drop, &herr);
@@ -129,7 +129,7 @@ int main(void)
     /* 注:drop 统计的是"队列满丢弃的发布尝试次数",发布方按契约重试,
      * 事件本体不丢——零丢失由下方 (tid,seq) 恰好一次证明,不要求 drop==0 */
     DG_CHECK(atomic_load(&s_payload_err) == 0); /* 载荷无串包 */
-    DG_CHECK(atomic_load(&s_received) == total);
+    DG_CHECK((uint32_t)atomic_load(&s_received) == total);
 
     /* 每条恰好一次:既不丢(==0)也不重(>=2) */
     int lost = 0, dup = 0;
